digitrecognizer/requirements.txt
actionlib==1.13.2
angles==1.9.13
pandas==1.2.4
apturl==0.5.2
asgiref==3.3.1
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
autopep8==1.5.7
backcall==0.2.0
bcrypt==3.1.7
blinker==1.4
bondpy==1.8.6
breezy==3.0.2
Brlapi==0.7.0
camera-calibration==1.15.3
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==0.4.23
catkin-pkg-modules==0.4.23
cbor==1.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
controller-manager==0.19.4
controller-manager-msgs==0.19.4
cryptography==2.8
cupshelpers==1.0
cv-bridge==1.15.0
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.7
diagnostic-analysis==1.10.4
diagnostic-common-diagnostics==1.10.4
diagnostic-updater==1.10.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.7
docutils==0.16
dulwich==0.19.15
duplicity==0.8.12.0
dynamic-reconfigure==1.7.1
empy==3.3.2
entrypoints==0.3
fasteners==0.14.1
fastimport==0.9.8
future==0.18.2
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gencpp==0.6.5
geneus==3.0.0
genlisp==0.4.18
genmsg==0.5.16
gennodejs==2.0.2
genpy==0.6.15
gpg===1.13.1-unknown
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
image-geometry==1.15.0
imageio==2.9.0
incremental==16.10.1
interactive-markers==1.12.0
ipykernel==5.5.0
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
joint-state-publisher==1.15.0
joint-state-publisher-gui==1.15.0
jupyter-client==6.1.12
jupyter-core==4.7.1
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
laser-geometry==1.6.7
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
matplotlib==3.4.1
message-filters==1.15.11
monotonic==1.5
mpi4py==3.0.3
netifaces==0.10.4
networkx==2.5.1
nose==1.3.7
numpy==1.20.2
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
paramiko==2.6.0
parso==0.8.1
pbr==5.6.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.18
protobuf==3.6.1
psutil==5.5.1
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycodestyle==2.7.0
pycryptodomex==3.6.1
pycups==1.9.73
pydot==1.4.1
PyGithub==1.43.7
Pygments==2.8.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-gitlab==2.0.1
python-gnupg==0.4.5
python-qt-binding==0.4.3
python-snappy==0.5.3
PyTrie==0.2
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.0.3
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
resource-retriever==1.12.6
roman==2.0.0
rosbag==1.15.11
rosboost-cfg==1.15.7
rosclean==1.15.7
roscreate==1.15.7
rosdep==0.20.1
rosdep-modules==0.20.1
rosdistro==0.8.3
rosdistro-modules==0.8.3
rosgraph==1.15.11
rosinstall==0.7.8
rosinstall-generator==0.1.22
roslaunch==1.15.11
roslib==1.15.7
roslint==0.12.0
roslz4==1.15.11
rosmake==1.15.7
rosmaster==1.15.11
rosmsg==1.15.11
rosnode==1.15.11
rosparam==1.15.11
rospkg==1.3.0
rospkg-modules==1.3.0
rospy==1.15.11
rosservice==1.15.11
rostest==1.15.11
rostopic==1.15.11
rosunit==1.15.7
roswtf==1.15.11
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.11
rqt-dep==0.4.11
rqt-graph==0.4.14
rqt-gui==0.5.2
rqt-gui-py==0.5.2
rqt-image-view==0.4.16
rqt-launch==0.4.9
rqt-logger-level==0.4.11
rqt-moveit==0.5.10
rqt-msg==0.4.10
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-publisher==0.4.10
rqt-py-common==0.5.2
rqt-py-console==0.4.10
rqt-reconfigure==0.5.4
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.13
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.9
rqt-rviz==0.6.1
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.2
rqt-top==0.4.10
rqt-topic==0.4.12
rqt-web==0.4.10
rviz==1.14.7
scikit-image==0.18.1
scipy==1.6.3
SecretStorage==2.3.1
sensor-msgs==1.13.1
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
sqlparse==0.4.1
systemd-python==234
testresources==2.0.1
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.5
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.5
tifffile==2021.4.8
toml==0.10.2
topic-tools==1.15.11
tornado==6.1
traitlets==5.0.5
Twisted==18.9.0
txaio==2.10.0
u-msgpack-python==2.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vcstools==0.1.42
wadllib==1.3.3
wcwidth==0.2.5
wrapt==1.11.2
wsaccel==0.6.2
wstool==0.1.18
xacro==1.14.6
xkit==0.0.0
zope.interface==4.7.1
